moved generate to fraudproofs. created types file
